requirements.txt
# Nexira / Ultimate AI System v8.0 - Python Dependencies
# Created by Xeeker & Claude - February 2026

# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
simple-websocket>=0.9.0

# AI/LLM
ollama>=0.1.6

# Document Processing (optional - gracefully skipped if missing)
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# Communication Services (Phase 3)
imapclient==2.3.1
requests==2.31.0
# twilio==8.10.0  # Uncomment when WhatsApp integration is needed

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# System Monitoring
psutil==5.9.6
# GPUtil==1.4.0  # Uncomment only if running on NVIDIA GPU hardware

# Utilities
python-dotenv==1.0.0
croniter==2.0.1
